WIP converting UploadSubmission and UploadStatus to work with the new Report model
